test(journal): move test packages around to reflect production structure
