+++ journal/src/test/java/io/camunda/zeebe/journal/JournalTest.java
@@ -1,33 +1,26 @@
 /*
- * Copyright © 2017 camunda services GmbH (info@camunda.com)
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
+ * Copyright Camunda Services GmbH and/or licensed to Camunda Services GmbH under
+ * one or more contributor license agreements. See the NOTICE file distributed
+ * with this work for additional information regarding copyright ownership.
+ * Licensed under the Zeebe Community License 1.1. You may not use this file
+ * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.journal.file;
+package io.camunda.zeebe.journal;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import io.camunda.zeebe.journal.Journal;
 import io.camunda.zeebe.journal.JournalException.InvalidChecksum;
 import io.camunda.zeebe.journal.JournalException.InvalidIndex;
-import io.camunda.zeebe.journal.JournalReader;
-import io.camunda.zeebe.journal.JournalRecord;
+import io.camunda.zeebe.journal.file.SegmentedJournal;
+import io.camunda.zeebe.journal.file.SegmentedJournalBuilder;
 import io.camunda.zeebe.journal.file.record.CorruptedLogException;
 import io.camunda.zeebe.journal.file.record.PersistedJournalRecord;
 import io.camunda.zeebe.journal.file.record.RecordData;
 import io.camunda.zeebe.journal.file.record.RecordMetadata;
-import io.camunda.zeebe.journal.file.util.TestJournalRecord;
+import io.camunda.zeebe.journal.util.LogCorrupter;
+import io.camunda.zeebe.journal.util.TestJournalRecord;
+import io.camunda.zeebe.util.buffer.BufferUtil;
 import java.io.File;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
@@ -600,13 +593,9 @@ class JournalTest {
     assertThat(reader.next()).isEqualTo(lastRecord);
   }
 
-  static PersistedJournalRecord copyRecord(final JournalRecord record) {
-    final DirectBuffer data = record.data();
-    final byte[] buffer = new byte[data.capacity()];
-    data.getBytes(0, buffer);
-
-    final UnsafeBuffer copiedData = new UnsafeBuffer(buffer);
-    final RecordData copiedRecord = new RecordData(record.index(), record.asqn(), copiedData);
+  private PersistedJournalRecord copyRecord(final JournalRecord record) {
+    final RecordData copiedRecord =
+        new RecordData(record.index(), record.asqn(), BufferUtil.cloneBuffer(record.data()));
 
     return new PersistedJournalRecord(
         new RecordMetadata(record.checksum(), copiedRecord.data().capacity()), copiedRecord);
