+++ journal/src/test/java/io/camunda/zeebe/journal/util/PosixPathAssert.java
@@ -5,7 +5,7 @@
  * Licensed under the Zeebe Community License 1.1. You may not use this file
  * except in compliance with the Zeebe Community License 1.1.
  */
-package io.camunda.zeebe.journal.file.util;
+package io.camunda.zeebe.journal.util;
 
 import java.io.File;
 import java.io.IOException;
